requirements.txt
numpy
PyYAML
flask
dash==0.19.0
dash-renderer==0.11.1
dash-html-components==0.8.0
dash-core-components==0.14.0
plotly
gunicorn
pandas
sqlalchemy
